backend/requirements.txt
# Backend API - Sistema Musuq Cloud

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
alembic>=1.13.0
bcrypt==4.0.1
